Fixes minor things

+++ src/pages/ManageSystem/ManageEmailList.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from 'react'
-import { useNavigate } from 'react-router-dom'
 import { toast } from 'react-toastify'
 import ButtonRectangle from '../../components/Buttons/ButtonRectangle'
 import TagInput from '../../components/TagInput'
@@ -14,10 +13,9 @@ import { isEmailValid } from '../../utils/validations'
 import './styles.css'
 
 function ManageEmailList() {
-  const navigate = useNavigate()
   const [emailTags, setEmailTags] = useState([])
 
-  const { emailLists } = useGetEmailLists()
+  const { emailLists, refresh: refreshEmailLists } = useGetEmailLists()
   const { createEmailList, deleteEmailList } = useAdminRoutes()
 
   const { state: emailListState } = useGetEmailListState()
@@ -43,12 +41,11 @@ function ManageEmailList() {
 
     if (newEmails.length > 0)
       createEmailList(newEmails, emailListState).then(async () => {
-        await deleteEmailList(deleteEmails)
-        navigate('/')
+        if (deleteEmails.length > 0) await deleteEmailList(deleteEmails)
+        refreshEmailLists()
       })
     else if (deleteEmails.length > 0)
-      deleteEmailList(deleteEmails).then(() => navigate('/'))
-    else navigate('/')
+      deleteEmailList(deleteEmails).then(() => refreshEmailLists())
   }
 
   useEffect(() => {
